refactor: 去除`log4js`
