+++ README.md
@@ -2,31 +2,94 @@
 
 ## 项目介绍
 
-`karin` 中译为「卡琳」，这个名字由[fuqiuluo](https://github.com/fuqiuluo)提供~
+`karin`（卡琳）是一款灵活、现代、极易扩展的 Node.js 插件化应用框架，专为开发者打造，助你轻松构建属于自己的高效工具链和自动化服务。
+
+✨ **主要特性**：
+- 插件化架构，支持热插拔，生态丰富
+- 一行命令即可初始化项目，快速上手
+- 丰富的 Web UI（基于 React + HeroUI），颜值与功能并存
+- 支持多种自动化场景、任务调度、依赖管理
+- 轻松集成第三方服务，打造属于你的"数字助理"
+- 社区活跃，持续更新，文档完善
+
+> 🦄 让开发变得像魔法一样有趣！
+
+## 🚀 稳定长期维护
+
+自 `1.8.0` 版本起，Karin 已进入**稳定长期维护阶段**。我们承诺持续修复 bug、优化体验，并欢迎社区力量共同完善生态。
 
 ## 快速开始
 
-[点击跳转文档](https://karin.fun/)
+[📚 查看最新文档](https://karin.fun/)
 
-当前文档可能存在滞后性，正在努力更新中。
+一键初始化：`pnpm create karin`
 
-有其他问题请加群讨论: 967068507
+> 当前文档可能存在滞后性，欢迎加入交流群（967068507）一起玩耍、提建议！
 
 ## 温馨提示
 
-当前`1.0.0`仍处于测试阶段，欢迎大家提出宝贵意见，我们会尽快修复。
+> Karin 现已稳定，放心食用！遇到问题欢迎提 Issue 或加群讨论，我们会持续优化。
+
+## 文档站说明
+
+我们提供多个文档站点供您访问，解决可能出现的访问困难：
+
+- **主文档站**: [https://karin.fun](https://karin.fun) (基于 GitHub Pages)
+- **镜像站点**:
+  - 憨憨镜像: [https://karin.hanhanz.top](https://karin.hanhanz.top) (**推荐访问**)
+  - Vercel 镜像: [https://docs.karin.fun](https://docs.karin.fun) (**推荐访问**)
+  - Deno 镜像: [https://karin.deno.dev](https://karin.deno.dev) (**推荐访问**)
+
+> 💡 主文档站托管在 GitHub 上，如访问不畅，推荐使用 Deno 镜像站
 
 ## 鸣谢
 
 - webui: [bietiaop](https://github.com/bietiaop)
 - docs: [ikenxuan](https://github.com/ikenxuan)
+- name: [fuqiuluo](https://github.com/fuqiuluo)
 
-> ps: 感谢两位大佬~
+> 🧙‍♂️ 感谢两位大佬的魔法加持！
 
 ### 贡献者
 
-> 🌟 星光闪烁，你们的智慧如同璀璨的夜空。感谢所有为**Karin**做出贡献的人！
+> 🌟 星光闪烁，你们的智慧如同璀璨的夜空。感谢所有为 **Karin** 做出贡献的人！
 
 [![贡献者](https://contributors-img.web.app/image?repo=KarinJS/Karin)](https://github.com/KarinJS/Karin/graphs/contributors)
 
 ![Alt](https://repobeats.axiom.co/api/embed/aaaa2759c8885691443a4d80e5753f975d4f250e.svg "Repobeats analytics image")
+
+---
+
+🎉 **加入我们，让 Karin 成为你开发路上的贴心伙伴！**
+
+## 常见问题
+
+- 文档没看懂？[点我提问](https://github.com/KarinJS/Karin/issues) 或加群 967068507
+- 插件不会写？欢迎参考[插件开发文档](https://karin.fun/plugins/)
+- 遇到 bug？大胆提 Issue，我们超快响应！
+
+## 如何参与贡献（PR）
+
+1. Fork 本仓库，创建你的分支
+2. 提交你的更改，附上简要说明
+3. 发起 Pull Request，耐心等待 Review
+4. 你的名字将出现在贡献者列表，收获一份开源荣誉！
+
+> 💡 欢迎任何形式的贡献，无论是代码、文档、建议还是灵感！
+
+## Issue 指南
+
+- 提交前请先搜索是否有类似问题
+- 尽量提供详细的复现步骤、环境信息和截图
+- 标题简明扼要，正文描述清晰
+- 遇到安全相关问题请私信维护者
+
+## 开源协议
+
+本项目基于 [MIT License](./LICENSE) 开源，欢迎自由使用、修改和分发。
+
+> 📢 记得给个 Star 支持我们，你的支持是我们最大的动力！
+
+## 更新日志
+
+我们定期发布更新，查看 [CHANGELOG](https://github.com/KarinJS/Karin/releases) 了解最新变化。
