+++ packages/core/src/utils/config/file/env.ts
@@ -9,6 +9,7 @@ import { listeners } from '@/core/internal/listeners'
 import type { Env } from '@/types/config/env'
 import { disconnectAll } from '@/adapter/onebot/connect'
 import { updateJwt } from '@/server/auth/jwt'
+import { LogMethodNames } from '@/types'
 
 /**
  * 自定义解析器
@@ -86,7 +87,7 @@ const initEnv = () => {
   watch<ReturnType<typeof parser>>(file, (old, data) => {
     dotenv.config({ path: file, override: true })
     process.env.RUNTIME = data.RUNTIME.value as 'node' | 'pm2' | 'tsx' || 'node'
-    logger.level = process.env.LOG_LEVEL || 'info'
+    logger.level = (process.env.LOG_LEVEL || 'info') as LogMethodNames
 
     if (old?.WS_SERVER_AUTH_KEY?.value !== data?.WS_SERVER_AUTH_KEY?.value) {
       logger.warn('[hmr] WebSocket服务器鉴权秘钥已更新')
